pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "multi-git-manager"
version = "1.0.0"
description = "Manage multiple Git accounts on a single system"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]

dependencies = [
    "click>=8.0.0",
    "pyyaml>=6.0",
    "rich>=12.0.0",
    "inquirer>=3.0.0",
]

[project.scripts]
mgit = "cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/multi-git-manager"
Issues = "https://github.com/yourusername/multi-git-manager/issues"